refactoring for entry points
