+++ edit_setup_init.py
@@ -1,20 +1,29 @@
 # coding: utf8
 import sys
-from ja_ginza import Japanese
 
 
 def main():
     edit_map = {
-        """    requirements = [parent_package + meta['spacy_version']]""":
-            """    requirements = [parent_package + "==" + (meta['spacy_version'][2:] if meta['spacy_version'].startswith('>=') else meta['spacy_version'])]""",
+        "    requirements = [parent_package + meta['spacy_version']]":
+            '    requirements = [parent_package + "==" + ('
+            "meta['spacy_version'][2:] if meta['spacy_version'].startswith('>=') else meta['spacy_version'])]",
         "        packages=[model_name],":
-            "        packages = [model_name, 'spacy.lang.{}','sudachipy'],".format(Japanese.lang),
+            "        packages = [model_name, "
+            "'ginza', "
+            "'sudachipy', "
+            "'sudachipy.dartsclone', "
+            "'sudachipy.dictionarylib', "
+            "'sudachipy.plugin', "
+            "'sudachipy.plugin.input_text', "
+            "'sudachipy.plugin.oov', "
+            "'sudachipy.plugin.path_rewrite'"
+            "],",
         "        package_data={model_name: list_files(model_dir)},":
             "        package_data={" +
             "model_name: list_files(model_dir), "
-            "'spacy.lang.{0}': "
-            "list_files(path.relpath('spacy/lang/{0}')) + "
-            "list_files(path.relpath('spacy/lang/{0}/resources')), "
+            "'ginza': "
+            "list_files(path.relpath('ginza')) + "
+            "list_files(path.relpath('ginza/resources')),"
             "'sudachipy': "
             "list_files(path.relpath('sudachipy')) + "
             "list_files(path.relpath('sudachipy/dartsclone')) + "
@@ -22,8 +31,12 @@ def main():
             "list_files(path.relpath('sudachipy/plugin')) + "
             "list_files(path.relpath('sudachipy/plugin/input_text')) + "
             "list_files(path.relpath('sudachipy/plugin/oov')) + "
-            "list_files(path.relpath('sudachipy/plugin/path_rewrite'))".format(Japanese.lang) +
-            "},"
+            "list_files(path.relpath('sudachipy/plugin/path_rewrite'))"
+            "},\n"
+            "        entry_points={"
+            '"spacy_factories": ["JapaneseCorrector = ginza:JapaneseCorrector"],'
+            '"spacy_languages": ["ja = ginza:Japanese"],'
+            "},",
     }
     setup_path = sys.argv[1]
     with open(setup_path, 'r') as f:
