+++ ginza/stop_words.py
