+++ ginza/syntax_iterators.py
