--- ja_ginza/resources/unk.def
@@ -1,35 +1,0 @@
-DEFAULT,5968,5968,3857,補助記号,一般,*,*,*,*
-SPACE,5966,5966,6056,空白,*,*,*,*,*
-KANJI,5139,5139,14657,名詞,普通名詞,一般,*,*,*
-KANJI,5129,5129,17308,名詞,普通名詞,サ変可能,*,*,*
-KANJI,4785,4785,18181,名詞,固有名詞,一般,*,*,*
-KANJI,4787,4787,18086,名詞,固有名詞,人名,一般,*,*
-KANJI,4791,4791,19198,名詞,固有名詞,地名,一般,*,*
-SYMBOL,5129,5129,17094,名詞,普通名詞,サ変可能,*,*,*
-NUMERIC,4794,4794,12450,名詞,数詞,*,*,*,*
-ALPHA,5139,5139,11633,名詞,普通名詞,一般,*,*,*
-ALPHA,4785,4785,13620,名詞,固有名詞,一般,*,*,*
-ALPHA,4787,4787,14228,名詞,固有名詞,人名,一般,*,*
-ALPHA,4791,4791,15793,名詞,固有名詞,地名,一般,*,*
-ALPHA,5687,5687,15246,感動詞,一般,*,*,*,*
-HIRAGANA,5139,5139,16012,名詞,普通名詞,一般,*,*,*
-HIRAGANA,5129,5129,20012,名詞,普通名詞,サ変可能,*,*,*
-HIRAGANA,4785,4785,18282,名詞,固有名詞,一般,*,*,*
-HIRAGANA,4787,4787,18269,名詞,固有名詞,人名,一般,*,*
-HIRAGANA,4791,4791,20474,名詞,固有名詞,地名,一般,*,*
-HIRAGANA,5687,5687,17786,感動詞,一般,*,*,*,*
-KATAKANA,5139,5139,10980,名詞,普通名詞,一般,*,*,*
-KATAKANA,5129,5129,14802,名詞,普通名詞,サ変可能,*,*,*
-KATAKANA,4785,4785,13451,名詞,固有名詞,一般,*,*,*
-KATAKANA,4787,4787,13759,名詞,固有名詞,人名,一般,*,*
-KATAKANA,4791,4791,14554,名詞,固有名詞,地名,一般,*,*
-KATAKANA,5687,5687,15272,感動詞,一般,*,*,*,*
-KANJINUMERIC,4794,4794,14170,名詞,数詞,*,*,*,*
-GREEK,5139,5139,11051,名詞,普通名詞,一般,*,*,*
-GREEK,4785,4785,13353,名詞,固有名詞,一般,*,*,*
-GREEK,4787,4787,13671,名詞,固有名詞,人名,一般,*,*
-GREEK,4791,4791,14862,名詞,固有名詞,地名,一般,*,*
-CYRILLIC,5139,5139,11140,名詞,普通名詞,一般,*,*,*
-CYRILLIC,4785,4785,13174,名詞,固有名詞,一般,*,*,*
-CYRILLIC,4787,4787,13495,名詞,固有名詞,人名,一般,*,*
-CYRILLIC,4791,4791,14700,名詞,固有名詞,地名,一般,*,*
